split cf api files
